dapp_builer_swarm/requirements.txt
-e /Users/tomisinjenrola/Documents/projects/blockchain/tuvana_labs/code/hive/hive-agent-py
#hive-agent @ git+https://github.com/hivenetwork-ai/hive-agent-py@main
aiosqlite==0.20.0
python-dotenv==1.0.1
requests==2.31.0
pydantic==2.7.1
